requirements.txt
# ============================================================
# Core dependencies
# ============================================================
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
speechbrain>=0.5.15

# Whisper — using OpenAI's package for clean detect_language API
openai-whisper>=20231117

# ============================================================
# Numerical / ML
# ============================================================
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0

# ============================================================
# Evaluation
# ============================================================
jiwer>=3.0.0

# ============================================================
# Config & Utilities
# ============================================================
pyyaml>=6.0
tqdm>=4.65.0

# ============================================================
# Visualization
# ============================================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# ============================================================
# Notebook support (Kaggle already has these)
# ============================================================
ipywidgets>=8.0.0
notebook>=7.0.0

# ============================================================
# NOTE: ffmpeg must be installed separately (system package)
# On Windows:  choco install ffmpeg  OR  download from ffmpeg.org
# On Kaggle:   pre-installed
# ============================================================
